pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["sahil <sahilofficial17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.151.0"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = "^0.32.0"
chromadb = "^0.5.18"
groq = "^0.11.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
